feat: tlsping and tcpping using step-by-step
This diff refactors the codebase to reimplement tlsping and tcpping
to use the step-by-step measurements style.

We're still missing a reference issue for now. We should first
take a look at this PoC and then create a reference issue.

We're still missing lots of tests for now. We should write those
tests before merging this pull request.

See docs/design/dd-003-step-by-step.md for more information on the
step-by-step measurement style.