fix: repair flaky test
Apparently, on macOS the context was always winning the race
while the same doesn't happen when running inside CI.

So, simplify the test and just basically make the first Write of
the TLS handshake fail, which should actually be fine as well.

Cc: @DecFox with whom I discussed this and other flaky tests
occurring in https://github.com/ooni/probe-cli/pull/815 today.