fix(oohelperd): use throw-away HTTPClient, Dialer, Resolver
This diff modifies the implementation of oohelperd in the release/3.15
branch to always use throw-away HTTPClient, Dialer, and Resolver.

The rationale of this change is to ensure we're not hitting limits of the
HTTPClient regarding the max number of connections per host.

This issue is described at https://github.com/ooni/probe/issues/2182.

While there, it feels more correct to use throw-away Dialer and Resolver.

While there, ensure we use the 8.8.8.8:53 resolver for the dialer and
the HTTPClient, which previously we were not doing 🤦.

We will need a different patch for `master` because there we have done
lots of refactoring of code using netx and netxlite (but the diff is going
to be more obvious to understand, so...)