refactor: Remove UnwrappedParallelResolver from trace
Since we are using the context for trace, we do not need a manual
wrapping step for our resolver. We can inject a trace directly into
the parallel resolver we require.

Suggestion(Simone): We may use a lambda function to pass the
resolver we need. This allows us to maintain a single ParallelResolver
which also handles mocking.