feat: support running measurements in parallel
This experimental diff stems from a discussion I had this morning
with @hellais, where this functionality seems helpful to him.

The diff itself breaks a bunch of places in the codebase, hence
it's not suitable for merging as is, but required more work, and
in particular, adapting tests and making sure we're not breaking
any of the supported use cases.

As an overview, the core of this diff consists of making sure
that the InputProcessor can run measurements in parallel.

In turn, this requires using a service for submitting and a service
for writing measurements back to disk. These two services should
be chained: saving to disk happens _after_ submitting such that we
write on disk a measurement with the report ID we got after we
submitted the measurement to the backend.

There is a queue between the measurer and the submitter with a buffer
of four places. I have experimentally determined that a larger than
this number would cause the code to really dishonour the maximum
runtime. In general, measuring could be faster than submitting thus
creating a large queue would mean a very long endgame condition
where we're busy submitting the queued measurements.

FWIW, if we're moving on with this approach, it would be quite
interesting to submit a compressed batch of measurements.

There are also some open questions about how it's best to report
errors, given that now the code runs asynchronously. We'll see.