Simplify the C API.
The general idea here is to avoid returning -1, having a more
uniform API, and allowing for sync calls.

It's a bit more work on both sides but it's more regular, so maybe
this change is for the greater good?

This will be a third pull request because it varies from
https://github.com/ooni/probe-cli/pull/850 mostly in
the way in which we interface at the FFI level.