feat: implement following redirects
We use a tracing style when following redirects. This reduces
the precision of the measurement but still collects useful data.

Additionally, we only use the system resolver for following
redirects, which is a bummer compared to using several resolvers.

The idea here is that we care mostly about the first URL.