fix: repair the Android build
Rather than hardcoding the NDK version inside a script, encode it
as a file in the filesystem, which is easier to share.

Call ./MOBILE/android/ensure in the github action so we fail if
we cannot install the desired Android NDK version.

Make sure we use the desired NDK by setting environment variables.

Use `-androidabi 21`, which:

1. is what rclone did: https://github.com/rclone/rclone/commit/8390ba4ca9f18fea546ecdfdf8f86aed5c36030f

2. is the minimum ABI used by probe-android: https://github.com/ooni/probe-android/blob/994651be52ff75ee3a00dba499ec7c7b40ab25c1/app/build.gradle#L10

Part of https://github.com/ooni/probe/issues/2130