feat(miniooni): use a remote miniooni instance to measure
This commit extends miniooni to be able to use a remote instance
of miniooni for performing OONI measurements.

The main use case for this functionality is to run measurements
inside another network that has censorship for QA purposes.

We currently support two transports for implementing remoting:

1. "tcp", which uses a cleartext TCP connection

2. "ssh", which uses SSH

The tcp transport is useful when you trust the network that transports
packets from and to the two miniooni instances.

The ssh transport covers the case where you want a secure channel
to transport packets between the two miniooni instances.

To start the remotetcp server, run this command as `root`:

```
./miniooni remotetcp
```

To start the remotessh server, run this command as `root`:

```
./miniooni remotessh
```

On the local side, you need to create `$HOME/.miniooni/remote/config.yaml`,
which must look like the following:

```YAML
remotes:
  foobar_tcp:
    address: "1.2.3.4:5555"
    transport: "tcp"

  foobar_ssh:
    address: "1.2.3.4:2222"
    transport: "ssh"
    ssh:
      user: "root"
```

The ports in the above example are the default ones used by
`remotetcp` and `remotessh`.

On the command line, you can enable remoting by adding `--remote=NAME`,
where NAME is the remote name. For example:

```
./miniooni -n --remote foobar_ssh example
```

Note that this command will only work if you have an active
instance of the ssh-agent (i.e., `SSH_AGENT_SOCK` is defined).

Having described the functionality, let us explain how we
implemented all of this.

We have reintroduced the possibility of completely taking over
the basic `netxlite` primitives:

1. dialing a TCP or UDP conn

2. looking up a domain name using getaddrinfo

3. creating a listening UDP socket

When you use `--remote NAME`, the code will do the following:

1. read the config file

2. figure out the right remote

3. establish a connection with the remote

4. create a TCP/IP stack in userspace

5. create a TUN device in userspace that gets all the packets
generated by the TCP/IP stack in userspace

6. overwrite netxlite primitives so they use the TCP/IP
stack in userspace instead of the Go standard library

7. setup routing between the connection with the remote and
the TUN device in userspace so we route packets

8. run the desired experiments as usual

On the server side, we do something similar, except that we
use a real TUN device as implemented by Linux. (We only
support Linux servers at the moment.)

Once the real TUN device has been created we setup masquerading
for it, and then we we route between the connection with the
miniooni client and the real TUN device.

We don't currently implement censorship, but the real TUN
device is clearly the right place where to do that.