refactor(webconnectivity@v0.5): improve logging clarity
We're bumping the experiment's version number because we changed the
name of the field used to contain late/duplicate DNS responses.

We have also changed the algorithm to determine #dnsDiff, however, the
change should only impact how we log this information.

Overall, here the idea is to provide users with a reasonably clear
explanation of how the probe maps observations to blocking and accessible.

Part of https://github.com/ooni/probe/issues/2237