feat: try all the available web connectivity THs
We introduce a fork of internal/httpx, named internal/httpapi, where
there is a clear split between the concept of an API endpoint (such as
https://0.th.ooni.org/) and of an API descriptor (such as using `GET`
to access /api/v1/test-list/url).

Additionally, httpapi allows to create a SequenceCaller that tries
to call a given API descriptor using multiple API endpoints. The
SequenceCaller will stop once an endpoint works or when all the available
endpoints have been tried unsuccessfully.

Then, we refactor web_connectivity@v0.4 and web_connectivity@v0.5 to
use a SequenceCaller for calling the web connectivity TH API.

Because this diff is designed to be backported to the release/3.16
branch, we have omitted additional changes to always use httpapi
where we are currently using httpx. Yet, to remind ourselves about
the need to do that, we have deprecated the httpx package.