refactor(probeservices): use internal testing (#994)
I was trying to (1) either move data types definitions inside probeservices
to internal/model or (2) make them private to reduce the amount of
exported definitions. However, I was severely limited in doing that
by the fact that probeservices uses external testing.

To overcome this issue, I'm switching to internal testing first and then
I will resume with the two points above, which are in line with my goal
for today of moving https://github.com/ooni/probe/issues/2372 forward.