refactor(probeservices): move data model inside internal/model (#995)
This diff refactors probeservices the move the full data model inside
of internal/model. Once I have all the model in there, I can re-introduce
code we sketched out in the past to keep the model in sync with the
API's swagger. Doing that without all the sources in the same directory
would have been quite a bit more difficult to do.

Part of https://github.com/ooni/probe/issues/2372