probeservices: sync up with v3.14.0 ooapi package
This diff ensures that the definitions we're using are in sync with
the model used by [v3.14.0's ooapi package](
https://github.com/ooni/probe-cli/tree/v3.14.0/internal/ooapi/apimodel),
which was an experiment at autogenerating APIs and cross checking
with ooni/api's swagger JSON file.

The only difference that remains pertains the name of the query
string parameter for /api/v1/test-list/urls, which must be plural
(as it is in the probeservices package) rather than singular (as
it is in the ooapi/apimodel package).

So, after merging this diff, we can say we have done part of the
work specified by https://github.com/ooni/probe/issues/2372. However,
we want to also check with the _current_ swagger.

(Because we already had a model that was consistent with the
swagger _at the time we wrote ooapi_, it seems reasonable for
us to start syncing up with that model, which is easy, and
then focus on importing the swagger and comparing with it.)