refactor(httpapi): logger belongs to endpoint
This diff changes httpapi Endpoint and Descriptor such that the
logger field is part of the Endpoint rather than Descriptor.

This changes makes sense because the logger to use should be the
same for every API call. What matters, instead, is whether to log
bodies, which remains a property of the API Descriptor.