refactor(httpapi): logger belongs to endpoint
This diff changes httpapi Endpoint and Descriptor such that the
logger field is part of the Endpoint rather than Descriptor.

This changes makes sense because the logger to use should be the
same for every API call. What matters, instead, is whether to log
bodies, which remains a property of the API Descriptor.

While I authored this diff originally as part of the work done
in https://github.com/ooni/probe-cli/pull/997, it seems it's
required in https://github.com/ooni/probe-cli/pull/1002 as well.

Hence, the reference issue is https://github.com/ooni/probe/issues/2379.