feat(httpapi): introduce the concept of Spec
The current httpapi code has the concept of Descriptor, which
describes how an API call should be made in terms of its request,
but which lacks a precise response definition.

Because our goal is to automatically generate swaggers from
the API spec, we need a more abstract way of describing an API.

This way will be the Spec. We will have two kind of Spec:

1. `SimpleSpec` is the specification for an API that returns
a raw response only consisting of bytes;

2. `TypedSpec[T]` is the specification for an API that
returns `*T` responses on success.

From a Spec, you can generate the `Descriptor`, so the previous
code does not change much.

Also, while there, recognize that, with a little tweak, we
don't need the SequenceCaller but we can do all the work
inside of the new APIs for calling specs.

Additionally, deprecate previous code that is now superseded
by this new approach--without removing this code for now.