fix(httpapi): we cannot actually dispose of SequenceCaller
The SequenceCaller concept is useful because potentially it can
incapsulate some state. Thus, in the future we may be able to
try the list of endpoints in the correct order.

So, it's actually worth keeping this concept and so we need
two distinct callers one for each possible calling style.

(I am starting to wonder whether this is a place where we could
use monads to decorate return values?)