cleanup(httpapi): remove optional logging-related complexity
1. we don't care _this much_ about body logging that we need to have
a specific test making sure we _actually_ log the bodies. If we break
body logging by incident, that's annoying but not really bad. In
exchange, we remove the most complex test of this package.

2. it's pointless to log that the content-type header is not the
expected one when expecting JSON. In case of parsing failure, the
user could use `-v` to turn on verbose mode and see logs. Hence,
we don't need to be this picky and this also allows us to zap some
extra code and testing complexity.

3. while there, avoid using `|foo|` in documentation comments
because the proper form is actually `[foo]`.

All this work is really just yak shaving to reduce the noise in a
diff for https://github.com/ooni/probe/issues/2372.